+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Changed
 
-* Drop Python 3.6 support
+- Switch to Werkzeug from uc-rfc6266-parser
+- Drop Python 3.6 support
 
 # [0.25.0] - 2022-07-19